+++ components/Dialogs/tests/unit/MDCAlertControllerTests.m
@@ -46,7 +46,25 @@ static NSDictionary<UIContentSizeCategory, NSNumber *> *CustomScalingCurve() {
 
 #pragma mark - Subclasses for testing
 
+/** A test fake for setting the @c traitCollection on the alert's view. */
+@interface MDCAlertControllerTestsFakeWindow : UIWindow
+
+/** Set to override the @c traitCollection property of the receiver. */
+@property(nonatomic, strong) UITraitCollection *traitCollectionOverride;
+@end
+
+@implementation MDCAlertControllerTestsFakeWindow
+
+- (UITraitCollection *)traitCollection {
+  return self.traitCollectionOverride ?: [super traitCollection];
+}
+
+@end
+
+/** A test fake to override the @c traitCollection of an @c MDCAlertController. */
 @interface MDCAlertControllerTestsControllerFake : MDCAlertController
+
+/** Set to override the value of @c traitCollection of the receiver. */
 @property(nonatomic, strong) UITraitCollection *traitCollectionOverride;
 @end
 
@@ -58,17 +76,22 @@ static NSDictionary<UIContentSizeCategory, NSNumber *> *CustomScalingCurve() {
 
 @end
 
+/** Expose private properties for testing. */
 @interface MDCAlertController (Testing)
 @property(nonatomic, nullable, weak) MDCAlertControllerView *alertView;
 @end
 
+/** Expose private properties for testing. */
 @interface MDCDialogPresentationController (Testing)
 @property(nonatomic) MDCDialogShadowedView *trackingView;
 @end
 
 #pragma mark - Tests
 
+/** Unit tests for @c MDCAlertController. */
 @interface MDCAlertControllerTests : XCTestCase
+
+/** The @c MDCAlertController being tested. */
 @property(nonatomic, nullable) MDCAlertController *alert;
 @end
 
@@ -640,6 +663,45 @@ than @c UIContentSizeCategoryLarge.
       XCTAssertNotEqualWithAccuracy(actionButton.titleLabel.font.pointSize, expectedPointSize,
                                     0.001);
     }
+  }
+}
+
+/**
+ Tests that setting a UIFontMetrics-based font will be updated based on changes to the
+ @c prefrredContentSizeCategory of the view's @c traitCollection.
+ */
+- (void)testAdjustsFontForContentSizeCategoryUpdatesFontWhenTraitCollectionChanges {
+  if (@available(iOS 11.0, *)) {
+    // Given
+    MDCAlertControllerTestsControllerFake *alert =
+        [[MDCAlertControllerTestsControllerFake alloc] init];
+    alert.title = @"Title";
+    alert.message = @"Message";
+    [alert addAction:[MDCAlertAction actionWithTitle:@"Action 1" handler:nil]];
+
+    // Prepare the Dynamic Type environment
+    UIFontMetrics *bodyMetrics = [UIFontMetrics metricsForTextStyle:UIFontTextStyleBody];
+    UITraitCollection *extraSmallTraits = [UITraitCollection
+        traitCollectionWithPreferredContentSizeCategory:UIContentSizeCategoryExtraSmall];
+    UIFont *originalFont = [UIFont fontWithName:@"Zapfino" size:20];
+    originalFont = [bodyMetrics scaledFontForFont:originalFont
+                    compatibleWithTraitCollection:extraSmallTraits];
+    alert.adjustsFontForContentSizeCategory = YES;
+    alert.titleFont = originalFont;
+    [alert loadViewIfNeeded];
+    MDCAlertControllerTestsFakeWindow *window = [[MDCAlertControllerTestsFakeWindow alloc] init];
+    window.traitCollectionOverride =
+        [UITraitCollection traitCollectionWithPreferredContentSizeCategory:
+                               UIContentSizeCategoryAccessibilityExtraExtraExtraLarge];
+    [window addSubview:alert.view];
+
+    // When
+    // Triggers UIFontMetrics-based fonts to resize within UILabel.
+    [alert.view layoutIfNeeded];
+
+    // Then
+    XCTAssertEqualObjects(alert.alertView.titleLabel.font.fontName, originalFont.fontName);
+    XCTAssertGreaterThan(alert.alertView.titleLabel.font.pointSize, originalFont.pointSize);
   }
 }