+++ components/TextFields/src/ContainedInputView/private/MDCTextControlLabelAnimation.h
@@ -25,12 +25,17 @@
 /**
  This method lays out the label in an animated fashion, often from normal position to the floating
  position, and vice versa.
+
+ This method does not attempt an animation and passes @c NO to the completion block if an animation
+ is already in progress. It does the animation and passes @c YES to the completion block if there is
+ no animation in progress.
  */
-+ (void)layOutLabel:(nonnull UILabel *)floatingLabel
++ (void)animateLabel:(nonnull UILabel *)label
                  state:(MDCTextControlLabelState)labelState
       normalLabelFrame:(CGRect)normalLabelFrame
     floatingLabelFrame:(CGRect)floatingLabelFrame
             normalFont:(nonnull UIFont *)normalFont
           floatingFont:(nonnull UIFont *)floatingFont
-     animationDuration:(NSTimeInterval)animationDuration;
+     animationDuration:(NSTimeInterval)animationDuration
+            completion:(void (^__nullable)(BOOL))completion;
 @end